+++ packages/cli/src/generators/microservice/index.ts
@@ -13,6 +13,7 @@ import {
   MIGRATION_CONNECTORS,
   SERVICES,
   BASESERVICEDSLIST,
+  BASESERVICECOMPONENTLIST,
 } from '../../enum';
 import {AnyObject, MicroserviceOptions} from '../../types';
 import {
@@ -291,6 +292,8 @@ export default class MicroserviceGenerator extends AppGenerator<MicroserviceOpti
     const type = this.options.facade ? 'facades' : 'services';
     if (!this.shouldExit()) {
       if (type === 'services') {
+        this.projectInfo.baseServiceComponentName =
+          this._setBaseServiceComponentName();
         const baseServiceDSList = this._setDataSourceName();
         this.projectInfo.baseServiceDSList = baseServiceDSList.filter(
           ds => ds.type === 'store',
@@ -337,7 +340,7 @@ export default class MicroserviceGenerator extends AppGenerator<MicroserviceOpti
       scripts[symlinkresolver] = symlinkresolver;
       scripts['resolve-links'] =
         'npm run symlink-resolver build ./node_modules/@local';
-      scripts['prestart'] = 'npm run rebuild && npm run openapi-spec';
+      scripts['prestart'] = 'npm run clean && npm run openapi-spec';
       scripts['rebuild'] = 'npm run clean && npm run build';
       scripts['start'] =
         'node -r ./dist/opentelemetry-registry.js -r source-map-support/register .';
@@ -368,7 +371,7 @@ export default class MicroserviceGenerator extends AppGenerator<MicroserviceOpti
       );
       this.spawnCommandSync('npm', ['i']);
       this.spawnCommandSync('npm', ['run', 'prettier:fix']);
-      this.destinationRoot(BACK_TO_ROOT);
+      this.destinationRoot(join(this.destinationPath(), BACK_TO_ROOT));
       if (this.options.migrations) {
         if (!this._migrationExists()) {
           this._createMigrationPackage();
@@ -421,6 +424,12 @@ export default class MicroserviceGenerator extends AppGenerator<MicroserviceOpti
     if (this.options.baseService) {
       return BASESERVICEDSLIST[this.options.baseService];
     } else return [];
+  }
+
+  private _setBaseServiceComponentName() {
+    if (this.options.baseService) {
+      return BASESERVICECOMPONENTLIST[this.options.baseService];
+    } else return undefined;
   }
 
   private _createDataSource() {
@@ -518,6 +527,30 @@ export default class MicroserviceGenerator extends AppGenerator<MicroserviceOpti
           destPath.replace('.ejs', ''),
       },
     );
+    // for the tpl files
+    const tplFilePath = this.templatePath(
+      join(MIGRATION_PACKAGE_TEMPLATE, 'packages', 'migrations'),
+    );
+    fs.readdirSync(
+      this.templatePath(
+        join(MIGRATION_PACKAGE_TEMPLATE, 'packages', 'migrations'),
+      ),
+    ).forEach(file => {
+      if (file.includes('.tpl')) {
+        const targetFileName = file.replace('.tpl', '');
+        const sourcePath = join(tplFilePath, file);
+        const destinationPath = join(
+          this.destinationRoot(),
+          MIGRATION_FOLDER,
+          targetFileName,
+        );
+        this.fs.copyTpl(sourcePath, destinationPath, {
+          name: this.options.name
+            ? this.options.name.toUpperCase()
+            : DEFAULT_NAME,
+        });
+      }
+    });
   }
 
   private _createCustomMigration() {
@@ -547,7 +580,7 @@ export default class MicroserviceGenerator extends AppGenerator<MicroserviceOpti
             sourceloopMigrationPath(this.options.baseService),
           ),
         ),
-        this.destinationPath(join(MIGRATION_FOLDER, name)),
+        this.destinationPath(join(MIGRATION_FOLDER, name, 'migrations')),
       );
       let connector = MIGRATION_CONNECTORS[DATASOURCES.POSTGRES]; // default
       if (this.options.datasourceType) {
@@ -590,6 +623,23 @@ export default class MicroserviceGenerator extends AppGenerator<MicroserviceOpti
         packageJsFile,
         JSON.stringify(packageJs, undefined, JSON_SPACING),
       );
+      const name = this.options.name
+        ? this.options.name.toUpperCase()
+        : DEFAULT_NAME;
+      let migEnv = this.fs.read(join(MIGRATION_FOLDER, '.env.example'));
+      const envToAdd = `\n${name}_DB_HOST= \n${name}_DB_PORT= \n${name}_DB_USER= 
+      \n${name}_DB_DATABASE= \n${name}_DB_PASSWORD=`;
+      migEnv = migEnv + envToAdd;
+      fs.writeFile(
+        join(MIGRATION_FOLDER, '.env.example'),
+        migEnv,
+        {
+          flag: 'w',
+        },
+        function () {
+          //This is intentional.
+        },
+      );
     } catch {
       //do nothing
     }
@@ -597,6 +647,7 @@ export default class MicroserviceGenerator extends AppGenerator<MicroserviceOpti
 
   async end() {
     if (this.projectInfo) {
+      this.spawnCommandSync('lerna', ['bootstrap', '--force-local']);
       this.projectInfo.outdir = this.options.name ?? DEFAULT_NAME;
     }
     await super.end();