+++ services/chat-service/src/repositories/index.ts
@@ -4,3 +4,4 @@
 // https://opensource.org/licenses/MIT
 export * from './message-recipient.repository';
 export * from './message.repository';
+export * from './attachment.repository';