+++ sandbox/in-mail-example/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.7.11 (2022-10-10)</small>
+
+**Note:** Version bump only for package @sourceloop/in-mail-example
+
+
+
+
+
+## <small>0.7.10 (2022-09-26)</small>
+
+**Note:** Version bump only for package @sourceloop/in-mail-example
+
+
+
+
+
 ## <small>0.7.9 (2022-09-21)</small>
 
 **Note:** Version bump only for package @sourceloop/in-mail-example