+++ sandbox/payment-example/backend/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.3.11 (2022-10-10)</small>
+
+**Note:** Version bump only for package @sourceloop/payment-example-backend
+
+
+
+
+
+## <small>0.3.10 (2022-09-26)</small>
+
+**Note:** Version bump only for package @sourceloop/payment-example-backend
+
+
+
+
+
 ## <small>0.3.9 (2022-09-21)</small>
 
 **Note:** Version bump only for package @sourceloop/payment-example-backend